refactor(planner): split the planner
