+++ crates/planner/src/builder/mod.rs
@@ -1,0 +1,6 @@
+// Re-export the main components from the builder module
+pub mod context;
+pub mod directive_handler;
+pub mod field_resolver;
+pub mod plan_builder;
+pub mod utils;
