+++ crates/planner/src/lib.rs
@@ -1,4 +1,4 @@
-#![forbid(unsafe_code)]
+#![allow(unsafe_code)]
 
 mod builder;
 mod plan;
@@ -6,7 +6,7 @@ mod request;
 mod response;
 mod types;
 
-pub use builder::PlanBuilder;
+pub use builder::plan_builder::PlanBuilder;
 pub use plan::{
     FetchNode,
     FlattenNode,
